+++ CHANGELOG.md
@@ -13,11 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
-## [1.1.0] - Upcoming
+## [1.1.0] - 2024.05.23
 
 ### Added
 
 - Added query builder.
+
+### Changed
+
+- Fixed bug in generating message when error is returned by Directus API.
 
 ## [1.0.0] - 2024.05.07